added member control

+++ resources/member.py
@@ -15,12 +15,23 @@ class MemberInGroup(MethodView):
     @blp.response(200, MemberSchema(many=True))
     def get(self, group_id):
         group = GroupModel.query.get_or_404(group_id)
+        current_user_id = get_jwt_identity()
+
+        if group.user_id != current_user_id:
+            abort(403, message="You are not authorized to view members of this group.")
+
         return group.members.all()
 
     @jwt_required()
     @blp.arguments(MemberSchema)
     @blp.response(201, MemberSchema)
     def post(self, tag_data, group_id):
+        current_user_id = get_jwt_identity()
+        group = GroupModel.query.get_or_404(group_id)
+
+        if group.user_id != current_user_id:
+            abort(403, message="You are not authorized to add members to this group.")
+
         member = MemberModel(group_id=group_id, **tag_data)
         try:
             db.session.add(member)
@@ -34,12 +45,15 @@ class MemberInGroup(MethodView):
 @blp.route("/transaction/<int:transaction_id>/member/<int:member_id>")
 class LinkMembersToTransaction(MethodView):
 
-    # @blp.arguments()
     @jwt_required()
     @blp.response(201, MemberSchema)
     def get(self, transaction_id, member_id):
         transaction = TransactionModel.get_or_404(transaction_id)
         member = MemberModel.get_or_404(member_id)
+        current_user_id = get_jwt_identity()
+
+        if transaction.group.user_id != current_user_id:
+            abort(403, message="You are not authorized to link members to this transaction.")
 
         transaction.members.append(member)
         try:
@@ -54,6 +68,10 @@ class LinkMembersToTransaction(MethodView):
     def delete(self, transaction_id, member_id):
         transaction = TransactionModel.get_or_404(transaction_id)
         member = MemberModel.get_or_404(member_id)
+        current_user_id = get_jwt_identity()
+
+        if transaction.group.user_id != current_user_id:
+            abort(403, message="You are not authorized to remove members from this transaction.")
 
         transaction.members.remove(member)
         try:
@@ -70,14 +88,25 @@ class Group(MethodView):
     @blp.response(200, MemberSchema)
     def get(self, member_id):
         member = MemberModel.query.get_or_404(member_id)
+        current_user_id = get_jwt_identity()
+
+        if member.group.user_id != current_user_id:
+            abort(403, message="You are not authorized to view this member.")
+
         return member
 
     @jwt_required()
     @blp.response(202, description="deletes a member if no transaction is tagged with it.")
     @blp.alt_response(404, description="member not found")
-    @blp.alt_response(400, description="Returns if the member is assigned to one or more transactions, the member is not deleted")
+    @blp.alt_response(400,
+                      description="Returns if the member is assigned to one or more transactions, the member is not deleted")
     def delete(self, member_id):
-        member = GroupModel.query.get_or_404(member_id)
+        member = MemberModel.query.get_or_404(member_id)
+        current_user_id = get_jwt_identity()
+
+        if member.group.user_id != current_user_id:
+            abort(403, message="You are not authorized to delete this member.")
+
         if not member.transactions:
             db.session.delete(member)
             db.session.commit()
